refactor(CurrentEditor): refactor editor components into CurrentEditor container
